requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.6.0
Django==3.1.6
django-allauth==0.44.0
django-ckeditor==6.0.0
django-filer==2.0.2
django-filter==2.4.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-polymorphic==3.0.0
django-ranged-response==0.2.0
django-registration-redux==2.9
django-simple-captcha==0.5.13
django.js==0.8.1
easy-thumbnails==2.7.1
filter==0.0.0.20200724
gunicorn==20.1.0
idna==2.10
mailgun==0.1.1
oauthlib==3.1.0
Pillow==8.1.0
pyactiveresource==1.0.1
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
Unidecode==1.1.2
urllib3==1.26.3
whitenoise==5.2.0
